[oauthclient] Rename package name to com.datastax.oss.kafka.oauth
