--- oauth-client/src/main/java/io/streamnative/pulsar/handlers/kop/security/oauth/OauthLoginCallbackHandler.java
@@ -1,85 +1,0 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.streamnative.pulsar.handlers.kop.security.oauth;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import javax.security.auth.callback.Callback;
-import javax.security.auth.callback.UnsupportedCallbackException;
-import javax.security.auth.login.AppConfigurationEntry;
-import org.apache.kafka.common.KafkaException;
-import org.apache.kafka.common.security.auth.AuthenticateCallbackHandler;
-import org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule;
-import org.apache.kafka.common.security.oauthbearer.OAuthBearerTokenCallback;
-
-/**
- * OAuth 2.0 login callback handler.
- */
-public class OauthLoginCallbackHandler implements AuthenticateCallbackHandler {
-
-    private ClientConfig clientConfig = null;
-
-    @Override
-    public void configure(Map<String, ?> configs, String saslMechanism, List<AppConfigurationEntry> jaasConfigEntries) {
-        if (!OAuthBearerLoginModule.OAUTHBEARER_MECHANISM.equals(saslMechanism)) {
-            throw new IllegalArgumentException("Unexpected SASL mechanism: " + saslMechanism);
-        }
-        if (Objects.requireNonNull(jaasConfigEntries).size() != 1 || jaasConfigEntries.get(0) == null) {
-            throw new IllegalArgumentException(
-                    String.format("Must supply exactly 1 non-null JAAS mechanism configuration (size was %d)",
-                            jaasConfigEntries.size()));
-        }
-
-        @SuppressWarnings("unchecked cast")
-        final Map<String, String> options = (Map<String, String>) jaasConfigEntries.get(0).getOptions();
-        if (options == null) {
-            throw new IllegalArgumentException("JAAS configuration options is null");
-        }
-        clientConfig = new ClientConfig(options);
-    }
-
-    @Override
-    public void close() {
-        // empty
-    }
-
-    @Override
-    public void handle(Callback[] callbacks) throws IOException, UnsupportedCallbackException {
-        if (clientConfig == null) {
-            throw new IllegalStateException("Callback handler not configured");
-        }
-        for (Callback callback : callbacks) {
-            if (callback instanceof OAuthBearerTokenCallback) {
-                try {
-                    handleCallback((OAuthBearerTokenCallback) callback);
-                } catch (KafkaException e) {
-                    throw new IOException(e.getMessage(), e);
-                }
-            } else {
-                throw new UnsupportedCallbackException(callback);
-            }
-        }
-    }
-
-    private void handleCallback(OAuthBearerTokenCallback callback) throws IOException {
-        if (callback.token() != null) {
-            throw new IllegalArgumentException("Callback had a token already");
-        }
-        try (final ClientCredentialsFlow flow = new ClientCredentialsFlow(clientConfig)) {
-            callback.token(flow.authenticate());
-        }
-    }
-}
